% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_raw_data}
\alias{add_raw_data}
\alias{add_raw_data.ModelWorkflow}
\title{Pulls data from local data source and adds to the data slot}
\usage{
add_raw_data(workflow, path)

\method{add_raw_data}{ModelWorkflow}(workflow, path)
}
\arguments{
\item{workflow}{and object of class ModelWorkflow}

\item{path}{data.frame or character string for the path to the object in the S3 bucket}
}
\value{
an object of class ModelWorkflow with updated data and metadata slots
}
\description{
Pulls data from local data source and adds to the data slot
}
