% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3.R
\name{add_raw_data_from_S3}
\alias{add_raw_data_from_S3}
\alias{add_raw_data_from_S3.ModelWorkflow}
\title{Pulls data from S3 and adds to the data slot}
\usage{
add_raw_data_from_S3(workflow, bucket, data_object_path, aws_credentials_path)

\method{add_raw_data_from_S3}{ModelWorkflow}(workflow, bucket, data_object_path,
  aws_credentials_path = "~/.aws/credentials")
}
\arguments{
\item{workflow}{and object of class ModelWorkflow}

\item{bucket}{character string for the name of the S3 bucket to pull}

\item{data_object_path}{character string for the path to the object in the S3 bucket}

\item{aws_credentials_path}{character string for the path to the aws credentials}
}
\value{
an object of class ModelWorkflow with updated data and metadata slots
}
\description{
Pulls data from S3 and adds to the data slot
}
