% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_audit.R
\name{audit}
\alias{audit}
\title{simple variable audit}
\usage{
audit(dat, variable)
}
\arguments{
\item{dat}{data.table - analytical data set containing the target and the explicative variables}

\item{variable}{a string or a number defining a column to run through audit}
}
\value{
The result is a 1-row data.table with 16 columns:
\itemize{
\item name_var - the name of the variable audited
\item varClass - the class of the variable audited
\item unique_values - number of unique values of the variable audited
\item pct_numeric - \% of rows consisting of only numbers or "-"
\item pct_na -  \% of NA values
\item pct_num_or_na - \% of NA values or rows rows consisting of only numbers or "-"
\item pct_null - \% of NULL values
\item pct_0 - \% of 0 values
\item list_values - first 5 unique values of the column
\item sums - sum
\item med - median
\item mea - mean
\item std - standard deviation
\item minimum - minimum
\item maximum - maximum
\item Flevels - "Levels of text variables (character or factor), "num" for numeric variables
}
}
\description{
This function allows you to calculate basic metrics on each variable of a dataset.
It evaluates one single variable at a time
}
\keyword{binomial}
\keyword{univariate}
