% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_audit.R
\name{audit_variables}
\alias{audit_variables}
\title{Runs a variable audit on all variables}
\usage{
audit_variables(workflow, variables, ...)
}
\arguments{
\item{workflow}{object of class ModelWorkflow}

\item{variables}{vector of variables}

\item{\dots}{arguments to be passed on to the audit function}
}
\description{
Runs a variable audit on all variables
}
