% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{auto_serialize}
\alias{auto_serialize}
\title{Serialize a model object to a character string if serialize_object is TRUE,
otherwise just return the object wrapped in a list}
\usage{
auto_serialize(object, serialize_object = FALSE, verbose = FALSE)
}
\arguments{
\item{object}{any R object}

\item{serialize_object}{logical should the object be serialised?}

\item{verbose}{logical print extra details on the serialized object

You wrap the object in a list so it can be stored in a single cell in a data frame}
}
\description{
Serialize a model object to a character string if serialize_object is TRUE,
otherwise just return the object wrapped in a list
}
