% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{bind_score_factory}
\alias{bind_score_factory}
\title{flattens a list of predictions output from a model factory into a data frame}
\usage{
bind_score_factory(predictions, dat, grouping_variables, target,
  cores = .MMoptions$cores)
}
\arguments{
\item{predictions}{list of predictions}

\item{dat}{data.frame containing the grouping factors and the target variable}

\item{grouping_variables}{character vector of grouping_variables}

\item{target}{name of target variable}

\item{cores}{number of processor cores}
}
\value{
data.frame of bound predictions
}
\description{
flattens a list of predictions output from a model factory into a data frame
}
