% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_feature_engineering.R
\name{build_model_matrix}
\alias{build_model_matrix}
\title{construct a matrix of the engineered features from the gam univariate analysis}
\usage{
build_model_matrix(dat, feature_name, univariate_gams)
}
\arguments{
\item{dat}{data.frame containing original features}

\item{feature_name}{character matching one of the columns in dat}

\item{univariate_gams}{data.frame containing the model information for the univariate gam}
}
\value{
data.frame
}
\description{
construct a matrix of the engineered features from the gam univariate analysis
}
