% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculate a series of metrics for model evaluation.}
\usage{
calc_metrics(observed, predicted, metrics = c("mape", "rsquared", "mase",
  "rmse", "mae"))
}
\arguments{
\item{observed}{a vector of observed values}

\item{predicted}{a vector of predicted values}

\item{metrics}{a character vector of metric names, metric functions named here need to be previously defined}
}
\description{
Calculate a series of metrics for model evaluation.
}
