% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_metadata}
\alias{create_metadata}
\title{Create Metadata}
\usage{
create_metadata(model_name, model_type, is_model_factory, use_case_name,
  use_case_start_dt, use_case_end_dt = NULL)
}
\arguments{
\item{model_name}{a character string with the name of the model}

\item{model_type}{a character string specifying the type of the model}

\item{is_model_factory}{logical}

\item{use_case_name}{a character string}

\item{use_case_start_dt}{a dttm string}

\item{use_case_end_dt}{a dttm string; Default is null}
}
\description{
Creates a default metadata file with initialization
}
\seealso{
Other metadata: \code{\link{load_metadata}},
  \code{\link{save_metadata}}
}
\keyword{internal}
