% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_model_workflow.R
\name{define_features}
\alias{define_features}
\title{Define the variables used as features and target in the model}
\usage{
define_features(workflow, feature_names, target_name, grouping_variables,
  fe_grouping_variables, ...)
}
\arguments{
\item{workflow}{an object of class ModelWorkflow}

\item{feature_names}{character vector}

\item{target_name}{character}

\item{grouping_variables}{character only used if model factory mode is set}

\item{fe_grouping_variables}{character vector used to define groups for feature engineering}

\item{\dots}{unused arguments}
}
\value{
an object of class ModelWorkflow

If the default '.' is supplied to `feature_names`, all variables except the target_name are used as features
}
\description{
Also defines grouping variables for model factory mode
}
