% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmfR.R
\docType{package}
\name{dmmfR}
\alias{dmmfR}
\alias{dmmfR-package}
\title{dmmfR: A package for simplifying and standardizing machine learning model workflows}
\description{
The dmmfR package provides S4 classes to define the elements of a ML workflow
and S3 functions for processing a workflow. Functions are provided to read from and write to AWS S3
}
