% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_data.ModelWorkflow}
\alias{drop_data.ModelWorkflow}
\title{Drops the data from the data slot}
\usage{
\method{drop_data}{ModelWorkflow}(workflow, drop = c("both", "raw",
  "testtrainvalid"))
}
\arguments{
\item{workflow}{object of class ModelWorkflow}

\item{drop}{character one of both, raw or testtrainvalid
export}
}
\description{
Drops the data from the data slot
}
