% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_data}
\alias{drop_data}
\title{Drops raw data and/or test/train/validation data}
\usage{
drop_data(workflow, drop)
}
\arguments{
\item{workflow}{an object of class ModelWorkFlow}

\item{drop}{character drop raw data, test/train/validation sets or all data}

\item{\dots}{passed through}
}
\value{
an object of class ModelWorkFlow
}
\description{
Saves memory and helper for creating child workflows
}
