% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_models}
\alias{drop_models}
\title{Drops models and/or model factory ensembles}
\usage{
drop_models(workflow, drop, ...)
}
\arguments{
\item{workflow}{an object of class ModelWorkFlow}

\item{drop}{character drop model, model factory or both}

\item{\dots}{passed through}
}
\value{
an object of class ModelWorkFlow
}
\description{
Saves memory and helper for creating child workflows
}
