% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_slots}
\alias{drop_slots}
\title{replaces any named slots in the workflow with new instantiations of the class}
\usage{
drop_slots(workflow, ...)
}
\arguments{
\item{workflow}{object of class ModelWorkflow}

\item{\dots}{character names of slots to drop}
}
\value{
ModelWorkflow object with slots dropped
}
\description{
replaces any named slots in the workflow with new instantiations of the class
}
