% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_feature_engineering.R
\name{feature_engineer_gam}
\alias{feature_engineer_gam}
\title{Runs a univariate gam on all features}
\usage{
feature_engineer_gam(workflow, variables, ...)
}
\arguments{
\item{workflow}{ModelWorkflow object}

\item{variables}{vector of variables}

\item{\dots}{arguments to be passed on to the audit function}
}
\description{
Runs a univariate gam on all features
}
