% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_feature_engineering.R
\name{generate_new_features_gam}
\alias{generate_new_features_gam}
\title{Wraps around build_model_matrix to construct a data frame of new engineered variables}
\usage{
generate_new_features_gam(feature_dat, variables, univariate_gams)
}
\arguments{
\item{feature_dat}{data.frame of variables to construct new features from}

\item{variables}{NULL or character vector of names of variables in feature_dat}

\item{univariate_gams}{data.frame}
}
\description{
Wraps around build_model_matrix to construct a data frame of new engineered variables
}
