% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_features}
\alias{get_features}
\title{Get Feature Names from the Metadata}
\usage{
get_features(workflow)
}
\arguments{
\item{workflow}{object of class ModelWorkflow}
}
\value{
an array of defined features names used for model training
}
\description{
Iterates through the feature list stored in the metadata of an oject of class `ModelWorkflow` and gets the feature names
}
