% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{get_S3_unserialise}
\alias{get_S3_unserialise}
\title{Pulls an object from S3 and tries to unserialise it}
\usage{
get_S3_unserialise(bucket, object,
  aws_credentials_path = "~/.aws/credentials")
}
\arguments{
\item{bucket}{AWS S3 bucket name}

\item{object}{file path inside the aws s3 bucket}

\item{aws_credentials_path}{a character string with the path where aws credentials are stored}
}
\description{
Designed to work with serialised model workflow objects
}
\examples{
\dontrun{
 wf <- get_S3_unserialise('dslab-aops-dev-dmmf', 
                          'models/unapproved/demand-forecasting-demo/test_workflow_serialised')
} 
}
\seealso{
unpack_object
}
