% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_metadata.R
\name{initialize_metadata}
\alias{initialize_metadata}
\title{Sets up initial metadata for a ModelWorkflow object}
\usage{
initialize_metadata(workflow, model_name, model_type, is_model_factory,
  use_case_name, use_case_start_dt, use_case_end_dt)
}
\arguments{
\item{workflow}{and object of class ModelWorkflow}

\item{model_name}{character string for the name of the workflow object}

\item{model_type}{character string}

\item{is_model_factory}{logical}

\item{use_case_name}{character string}

\item{use_case_start_dt}{Date}

\item{use_case_end_dt}{Date or NULL}
}
\value{
an object of class ModelWorkflow with an updated metadata slot
}
\description{
Sets up initial metadata for a ModelWorkflow object
}
