% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{live_predict}
\alias{live_predict}
\title{Run predictions based on the data in the MMWorkflow class object, return a dataframe of predictions}
\usage{
live_predict(workflow, score_set, prediction_fun = NULL,
  preprocess_fun = NULL, features = NULL, convert_to_dummy = FALSE,
  cores = 1, ...)
}
\arguments{
\item{workflow}{an object of class MMWorkflow}

\item{score_set}{data.frame to predict from}

\item{prediction_fun}{a function to produce predictions. Must include arguments for object and newdata or newx}

\item{preprocess_fun}{optional function to convert obs and pred data to the correct format in the scoring function}

\item{features}{character vector of features or NULL to select feature names from metadata}

\item{convert_to_dummy}{logical should the predictor dataset be converted to a model.matrix. Important for glmnet models}

\item{cores}{number of processor cores to run in model factory mode}

\item{\dots}{unused}
}
\value{
a data.frame of predictions
}
\description{
Run predictions based on the data in the MMWorkflow class object, return a dataframe of predictions
}
