% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_metadata}
\alias{load_metadata}
\title{Load Metadata}
\usage{
load_metadata(model_id, bucket,
  host = "http://ec2-52-215-64-233.eu-west-1.compute.amazonaws.com:8093")
}
\arguments{
\item{model_id}{a string}

\item{bucket}{a character string}

\item{host}{a url string of couchbase host; Default points to the couchbase hosted on kubernetes}
}
\description{
Load Metadata of a given model id from Couchbase
}
\seealso{
Other metadata: \code{\link{create_metadata}},
  \code{\link{save_metadata}}
}
\keyword{internal}
