% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{metrics_factory}
\alias{metrics_factory}
\title{Calculate metrics for data across grouping variables.}
\usage{
metrics_factory(predicted, dat, grouping_variables, target, predicted_var,
  metric_fun, cores = 1)
}
\arguments{
\item{predicted}{list of predictions with names of models}

\item{dat}{test set data}

\item{grouping_variables}{character vector of grouping variables}

\item{target}{character name of target}

\item{predicted_var}{character name of prediction column in prediction data}

\item{metric_fun}{function for evaluation metric calculation}

\item{cores}{number of processor cores to run the function in}
}
\description{
Calculate metrics for data across grouping variables.
}
