% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data.R
\docType{class}
\name{MMdata-class}
\alias{MMdata-class}
\alias{MMdata}
\title{An S4 class to represent the data component of a modelling workflow.}
\description{
An S4 class to represent the data component of a modelling workflow.
}
\section{Slots}{

\describe{
\item{\code{raw_data}}{data frame for the data as imported}

\item{\code{training_set}}{data frame for the data as imported}

\item{\code{test_set}}{data frame for the data as imported}

\item{\code{validation_set}}{data frame for the data as imported}
}}

