% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\docType{class}
\name{MMmodel-class}
\alias{MMmodel-class}
\alias{MMmodel}
\title{An S4 class to represent the model component of a modelling workflow.}
\description{
An S4 class to represent the model component of a modelling workflow.
}
\section{Slots}{

\describe{
\item{\code{model_object}}{representation of a model e.g. a train object}
}}

