% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model_factory.R
\docType{class}
\name{MMmodelFactory-class}
\alias{MMmodelFactory-class}
\alias{MMmodelFactory}
\title{An S4 class to represent the model factory component of a modelling workflow.}
\description{
Model factories are lists of models of the same type that run over multiple splits of the
dataset. e.g. over products and stores.
}
\section{Slots}{

\describe{
\item{\code{grouping_variables}}{character name of the  variable to split the data by}

\item{\code{model_object}}{representation of a model e.g. a train object}
}}

