% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_scoring.R
\docType{class}
\name{MMscoring-class}
\alias{MMscoring-class}
\alias{MMscoring}
\title{An S4 class to represent the scoring component of a modelling workflow.}
\description{
An S4 class to represent the scoring component of a modelling workflow.
}
\section{Slots}{

\describe{
\item{\code{test_set_predictions}}{data frame}

\item{\code{validation_set_predictions}}{data frame}

\item{\code{training_set_predictions}}{}
}}

