% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{model_factory}
\alias{model_factory}
\title{Runs multiple models of the same type over combinations of grouping variables}
\usage{
model_factory(dat, grouping_variables, features, target, formula = NULL,
  training_engine, feature_engineering = NULL,
  feature_engineering_dat = NULL, dummy = FALSE, cores = .MMoptions$cores,
  ts_model = FALSE, ...)
}
\arguments{
\item{dat}{a data frame}

\item{grouping_variables}{character vector of names of grouping variables}

\item{features}{character vector of grouping variable names}

\item{target}{character name of target variable}

\item{formula}{a model formula object e.g. for lmer models}

\item{training_engine}{function e.g. lmer, cv.glmnet, caret::train}

\item{cores}{number of processor cores to run the model factory on}

\item{ts_model}{logical use syntax for forecasting package models}

\item{\dots}{arguments to be passed on to the training engine}
}
\value{
list of objects of output class of the training engine
}
\description{
Calls train_single_model internally
}
\keyword{internal}
