% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model_workflow.R
\docType{class}
\name{ModelWorkflow-class}
\alias{ModelWorkflow-class}
\alias{ModelWorkflow}
\title{An S4 class to represent a modelling workflow.}
\description{
An S4 class to represent a modelling workflow.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{list}

\item{\code{data}}{object of class MMdata to contain the raw data, test and training sets}

\item{\code{features}}{list of predictor features and their classes}

\item{\code{model}}{object of class MMmodel}

\item{\code{model_factory}}{object of class MMmodelFactory}

\item{\code{scores}}{object of class MMscoring}

\item{\code{validation}}{list for validation}
}}

