% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{nuke_gam}
\alias{nuke_gam}
\title{A function to nuke GAM models}
\usage{
nuke_gam(model, verbose = FALSE)
}
\arguments{
\item{model}{GAM model object name you want to nuke}

\item{verbose}{logical print space saving details}
}
\description{
This function allows you to reduce the GAM model object size for big datasets
}
