% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{pack_object}
\alias{pack_object}
\title{A function to pack models in a character string}
\usage{
pack_object(object, verbose = FALSE)
}
\arguments{
\item{object}{object you want to serialize}

\item{verbose}{logical print extra details on the serialized object}
}
\description{
This function allows you to transform a model object to a character string
to be loaded into a sql db
}
\keyword{serialize}
