% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_execution.R
\name{parallel_bind}
\alias{parallel_bind}
\title{Generic function to run parallelised lapply and bind according to arbitrary binding functions}
\usage{
parallel_bind(dat, variables, fun, binding_fn = dplyr::bind_rows,
  cores = .MMoptions$cores, ...)
}
\arguments{
\item{dat}{a data frame}

\item{variables}{vector of variables to iterate over in lapply}

\item{fun}{function to apply at each iteration}

\item{binding_fn}{function to bind the individual list elements together with e.g. bind_rows}

\item{cores}{number of processor cores to assign to the task}

\item{\dots}{arguments to be passed to fun}
}
\value{
a dataframe or object returned from binding_fn
}
\description{
Generic function to run parallelised lapply and bind according to arbitrary binding functions
}
