% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_model_workflow.R
\name{partition_data}
\alias{partition_data}
\title{Partitions the data into a test, training and validation sets}
\usage{
partition_data(workflow, train_p, test_p, validation_p, audit_features)
}
\arguments{
\item{workflow}{object of class ModelWorkflow}

\item{train_p}{proportion of the data to go in the training set}

\item{test_p}{proportion of the data to go in the test set}

\item{validation_p}{proportion of the data to go in the training set}

\item{audit_features}{logical should the features be audited}
}
\value{
an object of class ModelWorkflow with an updated data slot
}
\description{
Partitions the data into a test, training and validation sets
}
