% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{predict_single_group}
\alias{predict_single_group}
\title{Generic prediction function}
\usage{
predict_single_group(model, dat, features, target, prediction_fun,
  feature_engineering, feature_engineering_dat, dummy, ts_model = FALSE, ...)
}
\arguments{
\item{model}{a model object}

\item{dat}{data frame for generating X for predictions}

\item{features}{character vector of feature names}

\item{target}{character name of target}

\item{prediction_fun}{function for predicting}

\item{feature_engineering}{engineering NULL or character}

\item{feature_engineering_dat}{data set for feature engineering models}

\item{dummy}{logical convert X to model.matrix}

\item{ts_model}{logical use syntax for forecasting package models}

\item{\dots}{passed to prediction_fn}
}
\description{
Generic prediction function
}
