% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{print.MMmodel}
\alias{print.MMmodel}
\title{print method for MMmodel object}
\usage{
\method{print}{MMmodel}(x, ...)
}
\arguments{
\item{x}{an object of class MMmodel}

\item{\dots}{arguments to be passed to the print method for the model object}
}
\description{
print method for MMmodel object
}
