% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model_factory.R
\name{print.MMmodelFactory}
\alias{print.MMmodelFactory}
\title{print method for MMmodelFactory object}
\usage{
\method{print}{MMmodelFactory}(x, n = 1, ...)
}
\arguments{
\item{x}{an object of class MMmodel}

\item{n}{which model in the factory should be printed}

\item{\dots}{arguments to be passed to the print method for the model object}
}
\description{
By default prints the first model in the factory
}
