% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{rsquared}
\alias{rsquared}
\title{Definition of rsquared error measure}
\usage{
rsquared(observed, predicted)
}
\arguments{
\item{observed}{a vector of observed values}

\item{predicted}{a vector of predicted values}
}
\description{
Definition of rsquared error measure
}
