% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv.R
\name{s3_get_csv}
\alias{s3_get_csv}
\title{AWS S3 Integration}
\usage{
s3_get_csv(bucket, object, aws_credentials_path = "~/.aws/credentials")
}
\arguments{
\item{bucket}{AWS S3 bucket name}

\item{object}{file path inside the aws s3 bucket}

\item{aws_credentials_path}{a character string with the path where aws credentials are stored}
}
\description{
Connects to S3 and returns an object of type csv
}
\keyword{internal}
