% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_pmml.R
\name{s3_put_pmml}
\alias{s3_put_pmml}
\title{Put PMML}
\usage{
s3_put_pmml(file, model_id, object_path, bucket,
  aws_credentials_path = "~/.aws/credentials")
}
\arguments{
\item{file}{a character string containing the filename (or full path) of the file you want to upload to S3.}

\item{model_id}{a character string containing the model id from the metadata}

\item{object_path}{a character string containing the path to store the oject inside an S3 bucket.}

\item{bucket}{a character string with the name of the bucket}

\item{aws_credentials_path}{a character string containing the path where the aws credentials are stored; by default it looks in the project root directory; see \code{\link{s3HTTP}} for more details}
}
\description{
Puts a pmml object into an S3 bucket
}
\keyword{internal}
