% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_metadata}
\alias{save_metadata}
\title{Save Model Metadata}
\usage{
save_metadata(metadata, bucket,
  host = "ec2-52-215-64-233.eu-west-1.compute.amazonaws.com:8091")
}
\arguments{
\item{metadata}{a metadata object (a list for now), created using \code{\link{create_metadata}} function}

\item{bucket}{a character string with the name of the bucket where the metadata should be stored}

\item{host}{a character string with the couchbase hostname}
}
\description{
Saves model metadata as a json document into a couchbasedb bucket. It mimicks the \code{cb.upsert} methon in python
}
\seealso{
Other metadata: \code{\link{create_metadata}},
  \code{\link{load_metadata}}
}
\keyword{internal}
