% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_defaultSummary}
\alias{score_defaultSummary}
\title{Very basic scoring function.}
\usage{
score_defaultSummary(obs, predictions, preprocess_fun = NULL)
}
\arguments{
\item{obs}{vector of observed values}

\item{predictions}{matrix of predicted values}

\item{preprocess_fun}{optional preprocessing function to coerce obs and predictions into a single data frame

Different predict methods produce different outputs. You can deal with this by defining a
custom preprocess_fun function}
}
\value{
list
}
\description{
implements defaultSummary from caret package
}
