% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_model}
\alias{score_model}
\title{Score the model based on the test or validation set based on the data in the MMWorkflow class object}
\usage{
score_model(workflow, score_set, print_scores, prediction_fun, scoring_fun,
  preprocess_fun, features, predictions_item_name, scores_item_name,
  convert_to_dummy, cores, ...)
}
\arguments{
\item{workflow}{an object of class MMWorkflow}

\item{score_set}{character - test_set or validation_set}

\item{print_scores}{logical flag to print the confusion matrix}

\item{prediction_fun}{a function to produce predictions. Must include arguments for object and newdata or newx}

\item{scoring_fun}{function to score the predictions}

\item{preprocess_fun}{optional function to convert obs and pred data to the correct format in the scoring function}

\item{features}{character vector of features or NULL to select feature names from metadata}

\item{predictions_item_name}{name of the list item the predictions are assigned to. Allows for multiple sets of predictions}

\item{scores_item_name}{name of the list item the scores are assigned to. Allows for multiple sets of predictions}

\item{convert_to_dummy}{logical should the predictor dataset be converted to a model.matrix. Important for glmnet models}

\item{cores}{number of processor cores to run in model factory mode}

\item{\dots}{unused



scores are assigned to either the test or validation slots in the scores slot}
}
\description{
Score the model based on the test or validation set based on the data in the MMWorkflow class object
}
