% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_audit.R
\name{select_audit_features}
\alias{select_audit_features}
\title{Use the audit to reduce the numbers of feartures}
\usage{
select_audit_features(audit, features)
}
\arguments{
\item{audit}{data.frame the audit table}

\item{features}{character vector of features}
}
\value{
character vector of audited features
}
\description{
More can be added here
}
