% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_feature_engineering.R
\name{select_univariate_gams}
\alias{select_univariate_gams}
\title{filter for the univariate models with the specified criteria}
\usage{
select_univariate_gams(univariate_gam)
}
\arguments{
\item{univariate_gam}{data.frame with feature engineering data

This does basic filtering for sane variables}
}
\value{
filtered data.frame
}
\description{
filter for the univariate models with the specified criteria
}
