% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_dummy}
\alias{to_dummy}
\title{convert all categorical variables to dummy vars}
\usage{
to_dummy(X)
}
\arguments{
\item{X}{a dataframe}
}
\value{
a matrix
}
\description{
convert all categorical variables to dummy vars
}
\keyword{internal}
