% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_model_workflow.R
\name{train_model}
\alias{train_model}
\title{Set up and train a model based on the data in the MMWorkflow class object}
\usage{
train_model(workflow, formula, training_engine, print_model, cores,
  convert_to_dummy, child, ...)
}
\arguments{
\item{workflow}{an object of class MMWorkflow}

\item{formula}{a model formula. This overrides the automatic feature selection}

\item{training_engine}{a function for model training}

\item{print_model}{logical should the model diagnostics be printed out}

\item{cores}{number of processor cores to use}

\item{convert_to_dummy}{logical should the predictor dataset be converted to a model.matrix. Important for glmnet models}

\item{child}{logical should the dataset be removed from the returned object}

\item{\dots}{arguments to be passed to the training engine

Examples of training engines are 'caret::train', 'randomForest', 'glmnet'. All training engines must have both
an x and y argument}
}
\description{
Set up and train a model based on the data in the MMWorkflow class object
}
