% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_single_model}
\alias{train_single_model}
\title{Generic function to produce a single model object}
\usage{
train_single_model(dat, features, target, formula = NULL, training_engine,
  feature_engineering = NULL, feature_engineering_dat = NULL,
  dummy = FALSE, ts_model = FALSE, ...)
}
\arguments{
\item{dat}{a data frame}

\item{features}{character vector of training feature names}

\item{target}{character target variable name}

\item{formula}{a model formuka object e.g. for lmer models}

\item{training_engine}{function e.g. lmer, cv.glmnet, caret::train}

\item{feature_engineering}{NULL or character name of the type of FE used}

\item{feature_engineering_dat}{NULL or data.frame of FE data}

\item{dummy}{logical transform X to dummy variable matrix}

\item{ts_model}{logical use syntax for forecasting package models}

\item{\dots}{arguments to be passed on to the training engine

Note that the formula interface cannot be used with feature engineering}
}
\value{
object of output class of the training engine
}
\description{
Works for either formula or x/y inputs
}
\details{
should work for caret, cv.glmnet, lmer, glmer etc.
}
\keyword{internal}
