% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_feature_engineering.R
\name{univariate_gam}
\alias{univariate_gam}
\title{univariate analysis for continuous target variables}
\usage{
univariate_gam(dat, predictor, target, nuke_mod = TRUE,
  serialize_model = .MMoptions$serialize_feature_engineering_models)
}
\arguments{
\item{dat}{data.frame containing the target and the explicative}

\item{predictor}{name of predictor variable}

\item{target}{character name of the target variable in dat}

\item{nuke_mod}{logical reduce size of the model}

\item{serialize_model}{serialize the model to a string}
}
\value{
data.frame of 1 row
}
\description{
This function allows you to identify the information potential of each explicative variable
}
