% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_metadata}
\alias{update_metadata}
\title{Generic function for updating metadata in a ModelWorkflow object
NOTE: should replace with recursive internal function for arbitrarily nested lists}
\usage{
update_metadata(workflow, ...)
}
\arguments{
\item{workflow}{object of class ModelWorkflow}

\item{\dots}{metadata name and value pairs}
}
\value{
object of class ModelWorkflow with updated metadata
}
\description{
Generic function for updating metadata in a ModelWorkflow object
NOTE: should replace with recursive internal function for arbitrarily nested lists
}
