% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_metadata.R
\name{write_metadata_to_store}
\alias{write_metadata_to_store}
\title{Upload the metadata slot from a ModelWorkflow object to the couchbase data store}
\usage{
write_metadata_to_store(workflow, cb_bucket, cb_host)
}
\arguments{
\item{workflow}{object of class ModelWorkflow}

\item{cb_bucket}{character name of the couchbase bucket}

\item{cb_host}{character hostname of the couchbase data store}
}
\value{
object of class ModelWorkflow passed through
}
\description{
S3 method forming part of a model workflow
}
