% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3.R
\name{write_model_to_S3}
\alias{write_model_to_S3}
\alias{write_model_to_S3.ModelWorkflow}
\title{Translate the model object to pmml format and then export to S3}
\usage{
write_model_to_S3(workflow, bucket, object_path, aws_credentials_path)

\method{write_model_to_S3}{ModelWorkflow}(workflow, bucket, object_path,
  aws_credentials_path = "~/.aws/credentials")
}
\arguments{
\item{workflow}{an object of class ModelWorkflow}

\item{bucket}{character name of the S3 bucket to write to}

\item{object_path}{character path to the location of the model store}

\item{aws_credentials_path}{character path to aws credentials}
}
\value{
an object of class ModelWorkflow with updated data and metadata slots
Models are saved in the object_path with `ModelID` (from the metadata) suffixed by 'pmml'
}
\description{
Translate the model object to pmml format and then export to S3
}
